apps/api/requirements.txt
# FastAPI and ASGI server
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
pydantic[email]>=1.10.7

# Database
prisma==0.9.1
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.13.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.29.0
tenacity==8.2.3

# AI and vector search
openai==1.3.9
numpy==1.24.3
scipy==1.11.4
tiktoken==0.5.2

# Testing
pytest==7.3.1
httpx==0.24.1
